fix(turbopack): edge api entry path should not have `/route` suffix
