+++ packages/next-swc/crates/next-core/src/middleware.rs
@@ -1,14 +1,10 @@
 use anyhow::Result;
 use indexmap::indexmap;
-use indoc::writedoc;
 use turbo_tasks::{Value, Vc};
-use turbo_tasks_fs::{rope::RopeBuilder, File, FileSystemPath};
-use turbopack_binding::turbopack::{
-    core::{
-        asset::AssetContent, context::AssetContext, module::Module, reference_type::ReferenceType,
-        virtual_source::VirtualSource,
-    },
-    ecmascript::utils::StringifyJs,
+use turbo_tasks_fs::{File, FileSystemPath};
+use turbopack_binding::turbopack::core::{
+    asset::AssetContent, context::AssetContext, module::Module, reference_type::ReferenceType,
+    virtual_source::VirtualSource,
 };
 
 use crate::util::{load_next_js_template, virtual_next_js_template_path};
@@ -54,39 +50,4 @@ pub async fn get_middleware_module(
     );
 
     Ok(module)
-}
-
-#[turbo_tasks::function]
-pub async fn wrap_edge_entry(
-    context: Vc<Box<dyn AssetContext>>,
-    project_root: Vc<FileSystemPath>,
-    entry: Vc<Box<dyn Module>>,
-    original_name: String,
-) -> Result<Vc<Box<dyn Module>>> {
-    use std::io::Write;
-    let mut source = RopeBuilder::default();
-    writedoc!(
-        source,
-        r#"
-            import * as module from "MODULE"
-
-            self._ENTRIES ||= {{}}
-            self._ENTRIES[{}] = module
-        "#,
-        StringifyJs(&format_args!("middleware_{}", original_name))
-    )?;
-    let file = File::from(source.build());
-    // TODO(alexkirsz) Figure out how to name this virtual asset.
-    let virtual_source = VirtualSource::new(
-        project_root.join("edge-wrapper.js".to_string()),
-        AssetContent::file(file.into()),
-    );
-    let inner_assets = indexmap! {
-        "MODULE".to_string() => entry
-    };
-
-    Ok(context.process(
-        Vc::upcast(virtual_source),
-        Value::new(ReferenceType::Internal(Vc::cell(inner_assets))),
-    ))
 }
