+++ packages/next-swc/crates/next-core/src/next_edge/mod.rs
@@ -1,4 +1,5 @@
 pub mod context;
+pub mod entry;
 pub mod page_transition;
 pub mod route_regex;
 pub mod route_transition;
